Added support for multiple sets of settings via the --profile switch to module-starter command.
To support profiles rather than extending the homegrown format of the
config file further, I standarized on an INI file, however it is hard
to read both INI and the current format from the same file, so profile
support is only available if you create a config.ini file, which is
favoured over the existing "config" file.

When a (valid) profile is specified the configuration is built from
the default config values, which are overridden by the settings in
the specified profile, but which are overridden by the command line
options.

The reason why this was added was to allow you to create modules and/or
templates for specific solutions. For example you might want a specific
installation when using Dancer or Catalyst. Similarly you might want one
installation in which you use Dancer with a Bootstrap/jQuery/Nginx
configuration. You could therefore create a template with that
exact configuration each time that you wanted to start.