Add support for multiple authors to Module::Starter; Issue #25
The included changes add support for multiple authors to Module::Starter and propogate to the chosen builder option. To do this, it seemed appropriate to remove the --email option and ask that users' specify both the author name and email as a single option, which most of the builders expect anyway:

    module-starter --module=Foo::Bar,Foo::Bat \
        --author="Andy Lester <andy@petdance.com> \
        --author="Sawyer X <sawyerx@cpan.org>

The attribute that previously held the author, `$self->{author}` now holds an arrayref of authoremails. Which are passed in turn to Module::Build and ExtUtils::MakeMaker as arrayrefs and to Module::Install as a string.

I tried to follow the existing conventions that were used within the module, but please let me know if I missed anything and/or there are other change you would like me to make.